Removed duplication between archstrike-iso and archiso by making archiso a submodule. In doing so, configs/minimal was moved to configs/minimal-deprecated for convenient reference until functionality is verified.
Since this commit is larger than it should be, here is a breakdown of
the changes committed:
    - added archiso as submodule
    - added bin/port-archiso-releng to keep in sync with upstream
      changes... that is, it copies and updates everything we need
    - moved minimal to minimal-deprecated... once it all works,
      minimal-deprecated can be removed
    - merged the profile from minimal with releng

The best documentation on the differences between the archiso releng
profile and our archstrike profile is the script
bin/port-archiso-releng. Each function in port-archiso-releng is named
in a self-documenting way and encapsulates all changes made to releng.